command: Split source MANIFEST generation into its own command
This will avoid the generation of MANIFEST when building a Wheel
